db, controller, pop up forms, general UI changes and the icon changed
in db and controller ı had to add the serving_Size column since liquids are in ml, or liters.(it doesnt make any sense if u say 100gram of milk). other than that,  others category will be appearing on the bottom of the page. Also in the selectingredients, if the ingredient u want to add is in "other" category, you are going to select or add new category in the next page of the forum.